requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.1
Click==7.0
dnspython==1.16.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
jaraco.functools==2.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
portend==2.4
pycryptodome==3.8.1
PyJWT==1.7.1
pymongo[tls,srv]==3.6.1
python-dotenv==0.10.2
pytz==2018.9
pywin32==224
requests==2.21.0
six==1.12.0
tempora==1.14
urllib3==1.24.1
Werkzeug==0.15.2
zc.lockfile==1.4
